requirements.txt
langchain
pypdf
python-dotenv
pinecone-client
deeplake
openai
tiktoken
transformers
deeplake
sentencepiece
gpt4all
pyvis
apify-client